pyproject.toml
[project]
name = "discoursevqa-inference-llm-as-a-judge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.17",
    "moviepy>=2.2.1",
    "openai>=1.101.0",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.1.1",
]
